requirements.txt
cffi==1.15.1
cryptography==40.0.1
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
